refactor: fixed grammar error in package name
